Move audioEffectManager to media package
